refactor: folder for block templates
