+++ frontend/src/utils/blockTemplate/familyTemplates.ts
@@ -1,158 +1,16 @@
 import type { BlockOptions, BlockStyleMap, Slot } from "@/types";
 
-function getBlockTemplate(
-	type:
-		| "body"
-		| "container"
-		| "fit-container"
-		| "header"
-		| "list"
-		| "settings-dialog"
-		| "fallback-component"
-		| "empty-component"
-		| "missing-component"
-): BlockOptions {
-	switch (type) {
-		case "body":
-			return {
-				componentId: "root",
-				componentName: "div",
-				blockName: "body",
-				originalElement: "body",
-				children: [],
-				baseStyles: {
-					display: "flex",
-					flexDirection: "row",
-					flexShrink: 0,
-					width: "inherit",
-					overflowX: "hidden",
-					height: "100%",
-				}
-			};
-
-		case "container":
-			return {
-				componentName: "container",
-				originalElement: "div",
-				blockName: "container",
-				baseStyles: {
-					display: "flex",
-					flexDirection: "row",
-					flexShrink: 1,
-				} as BlockStyleMap,
-			};
-
-		case "fit-container":
-			return {
-				componentName: "container",
-				originalElement: "div",
-				blockName: "container",
-				baseStyles: {
-					display: "flex",
-					flexDirection: "row",
-					flexShrink: 1,
-					height: "fit-content",
-					width: "fit-content",
-				} as BlockStyleMap,
-			};
-
-		case "header":
-			return {
-				componentName: "header",
-				blockName: "header",
-				originalElement: "header",
-				baseStyles: {
-					display: "flex",
-					flexDirection: "row",
-					width: "100%",
-					height: "fit-content",
-					padding: "10px 12px",
-					backgroundColor: "var(--surface-base)",
-					borderStyle: "solid",
-					borderWidth: "0px 0px 1px 0px",
-					borderColor: "var(--outline-gray-1)",
-				} as BlockStyleMap,
-				children: [
-					{
-						componentName: "Breadcrumbs",
-						componentProps: {
-							items: [
-								{
-									label: "Home",
-									route: { name: "Home" },
-								},
-								{
-									label: "List",
-									route: "/components/breadcrumbs",
-								},
-							],
-						}
-					},
-					{
-						componentName: "Button",
-						componentProps: {
-							label: "Create",
-							iconLeft: "plus",
-							variant: "solid",
-						},
-						baseStyles: {
-							marginLeft: "auto",
-						} as BlockStyleMap,
-					}
-				],
-			}
-
-		case "list":
-			return listTemplate();
-
-		case "settings-dialog":
-			return settingsDialogTemplate();
-
-		case "fallback-component":
-			return {
-				componentName: "p",
-				originalElement: "__raw_html__",
-				innerHTML: `<div style="color: red;background: #f4f4f4;display:flex;flex-direction:column;position:static;top:auto;left:auto;width: 600px;height: 275px;align-items:center;font-size: 30px;justify-content:center"><p>Component missing</p></div>`,
-				baseStyles: {
-					height: "fit-content",
-					width: "fit-content",
-				}
-			}
-
-		case "empty-component":
-			return {
-				componentName: "container",
-				originalElement: "div",
-				baseStyles: {
-					height: "200px",
-					width: "100%",
-				} as BlockStyleMap,
-			};
-
-		case "missing-component":
-			return {
-				componentName: "HTML",
-				originalElement: "__raw_html__",
-				innerHTML: `<div style="color:#E86C13;background:#F8F8F8;display:flex;width:300px;height:150px;align-items:center;font-size:16px;justify-content:center"><p>Component Missing</p></div>`,
-				baseStyles: {
-					height: "fit-content",
-					width: "fit-content",
-				} as BlockStyleMap,
-			};
-	}
-}
-
-// --- Compound component templates -------------------------------------------
-// The list and settings families are compositional: one drop should yield a
-// whole, working tree. Content nests through `children` (which Studio renders
-// into each component's default slot) — the one exception is ListRows: its row
-// template must live in a real slot, because Studio injects the scoped-slot props
-// (item/index/value) only into slot content, not children. Everything below
-// ListRows inherits that scope ambiently, so it can stay as plain children.
+// Block templates for component families (List, Settings Dialog). Families are
+// compositional: one drop of the primary should yield a whole, working tree.
+// Content nests through `children` (which Studio renders into each component's
+// default slot) — the one exception is ListRows: its row template must live in a
+// real slot, because Studio injects the scoped-slot props (item/index/value) only
+// into slot content, not children. Everything below ListRows inherits that scope
+// ambiently, so it can stay as plain children.
 
 // A List seeded in column mode: a two-column header plus ListRows bound to sample
 // items, each row reading `item`/`value` from the scope ListRows exposes.
-function listTemplate(): BlockOptions {
+export function listTemplate(): BlockOptions {
 	return {
 		componentName: "List",
 		blockName: "List",
@@ -194,7 +52,7 @@ function listTemplate(): BlockOptions {
 // (a teleported modal), so it's edited on a fragment canvas (editInFragmentMode) via
 // ProxySettingsDialog, which renders the inner TabsRoot inline. `shortcut` and
 // `unmountOnHide` are turned off here because both are hostile to the editor.
-function settingsDialogTemplate(): BlockOptions {
+export function settingsDialogTemplate(): BlockOptions {
 	return {
 		componentName: "SettingsDialog",
 		blockName: "Settings Dialog",
@@ -300,5 +158,3 @@ function withDefaultSlot(content: BlockOptions[]): Record<string, Slot> {
 		default: { slotName: "default", slotContent: content },
 	} as unknown as Record<string, Slot>;
 }
-
-export default getBlockTemplate;
